13/06/2023

+++ MISA.WebFresher2023.Demo.BL/Service/BaseService.cs
@@ -3,9 +3,11 @@ using DocumentFormat.OpenXml.Office2010.Excel;
 using MISA.WebFresher2023.Demo.BL.Dto;
 using MISA.WebFresher2023.Demo.Common;
 using MISA.WebFresher2023.Demo.Common.Attribute;
+using MISA.WebFresher2023.Demo.Common.Constant;
 using MISA.WebFresher2023.Demo.Common.MyException;
 using MISA.WebFresher2023.Demo.DL.Repository;
 using System.ComponentModel.DataAnnotations;
+using System.Reflection;
 
 namespace MISA.WebFresher2023.Demo.BL.Service
 {
@@ -43,7 +45,11 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         /// Author: LeDucTiep (23/05/2023)
         public virtual async Task<int> DeleteAsync(Guid id)
         {
-            DeleteValidate(id);
+            // Kiểm tra lỗi
+            List<int> errorCodes = await DeleteValidate(id);
+            if (errorCodes.Any())
+                throw new BadRequestException(errorCodes);
+
             /// Xóa 
             return await _baseRepository.DeleteAsync(id);
         }
@@ -65,17 +71,21 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         /// Hàm lấy một bản ghi
         /// </summary>
         /// <param name="id">Id của bản ghi cần lấy </param>
-        /// <returns>Task<TEmployeeId></returns>
+        /// <returns>TEntityDto?</returns>
         /// Author: LeDucTiep (23/05/2023)
         public virtual async Task<TEntityDto?> GetAsync(Guid id)
         {
-            var entity = await _baseRepository.GetAsync(id);
-
-            if (entity == null)
+            // Kiểm tra lỗi
+            if (id.Equals(Guid.Empty))
             {
-                return default;
+                List<int> errorList = new() { (int)EmployeeErrorCode.GuidInvalid };
+                throw new BadRequestException(errorList);
             }
 
+            // Thực hiện
+            var entity = await _baseRepository.GetAsync(id);
+
+            // Trả về
             var entityDto = _mapper.Map<TEntityDto>(entity);
 
             return entityDto;
@@ -91,13 +101,28 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         {
             TEntityDto entityDto = _mapper.Map<TEntityDto>(entity);
 
-            Validate(entityDto);
+            List<int> errorCodes = new();
 
-            PostValidate(entityDto);
+            errorCodes.AddRange(Validate(entityDto));
+
+            try
+            {
+                List<int> ints = await PostValidate(entityDto);
+
+                errorCodes.AddRange(ints);
+            }
+            catch { }
+
+
+            if (errorCodes.Any())
+                throw new BadRequestException(errorCodes);
 
             TEntity entity1 = _mapper.Map<TEntity>(entity);
 
-            await _baseRepository.PostAsync(entity1);
+            int changedCount = await _baseRepository.PostAsync(entity1);
+
+            if (changedCount == 0)
+                throw new InternalException();
 
             TEntityDto entity2 = _mapper.Map<TEntityDto>(entity1);
 
@@ -114,9 +139,18 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         {
             TEntityDto entityDto = _mapper.Map<TEntityDto>(entity);
 
-            Validate(entityDto);
+            List<int> errorCodes = new();
 
-            UpdateValidate(id, entityDto);
+            errorCodes.AddRange(Validate(entityDto));
+
+            try
+            {
+            errorCodes.AddRange(await UpdateValidate(id, entityDto));
+            }
+            catch { }
+
+            if (errorCodes.Any())
+                throw new BadRequestException(errorCodes);
 
             // Thêm trường id để trả về
             TEntity _entity = _mapper.Map<TEntity>(entity);
@@ -130,7 +164,7 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         /// </summary>
         /// <returns>Danh sách bản ghi</returns>
         /// Author: LeDucTiep (23/05/2023)
-        public async Task<IEnumerable<TEntityDto>> GetAllAsync()
+        public virtual async Task<IEnumerable<TEntityDto>> GetAllAsync()
         {
             // Gọi đến procedure
             IEnumerable<TEntity> myList = await _baseRepository.GetAllAsync();
@@ -142,7 +176,7 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         /// Hàm validate
         /// </summary>
         /// Author: LeDucTiep (09/06/2023)
-        public void Validate(TEntityDto entity)
+        public virtual List<int> Validate(TEntityDto entity)
         {
             System.Reflection.PropertyInfo[] properties = typeof(TEntityDto).GetProperties();
 
@@ -152,25 +186,55 @@ namespace MISA.WebFresher2023.Demo.BL.Service
             {
                 var value = property.GetValue(entity, null);
 
+
+                // Xét bắt buộc 
                 var attributeRequired = (MSRequiredAttribute?)property.GetCustomAttributes(typeof(MSRequiredAttribute), false).FirstOrDefault();
 
-                if (attributeRequired != null && string.IsNullOrEmpty(value.ToString()))
+                if (attributeRequired != null)
                 {
-                    errorCodes.Add(attributeRequired.ErrorCode);
+                    if (value == null)
+                        errorCodes.Add(attributeRequired.ErrorCode);
+                    else if (property.PropertyType == typeof(Guid) && (Guid)value == Guid.Empty)
+                        errorCodes.Add(attributeRequired.ErrorCode);
+                    else if (property.PropertyType == typeof(string) && (string)value == string.Empty)
+                        errorCodes.Add(attributeRequired.ErrorCode);
+
                 }
 
+                // Xét độ dài 
                 var attributeMaxLength = (MSMaxLengthAttribute?)property.GetCustomAttributes(typeof(MSMaxLengthAttribute), false).FirstOrDefault();
 
-                if (attributeMaxLength != null && value != null && value.ToString().Length > attributeMaxLength.Length)
+                if (attributeMaxLength != null && value != null)
                 {
-                    errorCodes.Add(attributeMaxLength.ErrorCode);
+                    int valueLength = value.ToString().Length;
+                    int maxLength = attributeMaxLength.Length;
+                    bool isTooLong = valueLength > maxLength;
+                    if (isTooLong)
+                    {
+                        errorCodes.Add(attributeMaxLength.ErrorCode);
+                    }
+                }
+
+
+                // Xét ngày tháng
+                var validDateInThePast = (MSValidDateInThePastAttribute?)property.GetCustomAttributes(typeof(MSValidDateInThePastAttribute), false).FirstOrDefault();
+
+                if (validDateInThePast != null && value != null && !MSValidDateInThePastAttribute.IsValid(value))
+                {
+                    errorCodes.Add(validDateInThePast.ErrorCode);
+                }
+
+
+                // Xét email
+                var emailAttribute = (MSEmailAttribute?)property.GetCustomAttributes(typeof(MSEmailAttribute), false).FirstOrDefault();
+
+                if (emailAttribute != null && !string.IsNullOrEmpty((string)value) && !MSEmailAttribute.IsValid(value))
+                {
+                    errorCodes.Add(emailAttribute.ErrorCode);
                 }
             }
 
-            if (errorCodes.Count > 0)
-            {
-                throw new BadRequestException(errorCodes);
-            }
+            return errorCodes;
         }
 
         /// <summary>
@@ -179,8 +243,14 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         /// <typeparam name="T">Thực thể</typeparam>
         /// <param name="entity">Thực thể</param>
         /// Author: LeDucTiep (09/06/2023)
-        public virtual void PostValidate(TEntityDto entity)
+        public virtual async Task<List<int>> PostValidate(TEntityDto entity)
         {
+            Task<List<int>> task = new(
+                    () => new List<int>()
+                );
+            task.Start();
+
+            return await task;
         }
 
         /// <summary>
@@ -190,8 +260,14 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         /// <param name="entity">Thực thể</param>
         /// <param name="id">Id của bản ghi</param>
         /// Author: LeDucTiep (09/06/2023)
-        public virtual void UpdateValidate(Guid id, TEntityDto entity)
+        public virtual async Task<List<int>> UpdateValidate(Guid id, TEntityDto entity)
         {
+            Task<List<int>> task = new(
+                    () => new List<int>()
+                );
+            task.Start();
+
+            return await task;
         }
 
         /// <summary>
@@ -199,8 +275,14 @@ namespace MISA.WebFresher2023.Demo.BL.Service
         /// </summary>
         /// <param name="id">Id của bản ghi</param>
         /// Author: LeDucTiep (09/06/2023)
-        public virtual void DeleteValidate(Guid id)
+        public virtual async Task<List<int>> DeleteValidate(Guid id)
         {
+            Task<List<int>> task = new(
+                    () => new List<int>()
+                );
+            task.Start();
+
+            return await task;
         }
 
         #endregion
